Refactor profile_screen.dart into separate files
